+++ src/vidore_benchmark/retrievers/base_vision_retriever.py
@@ -57,7 +57,7 @@ class BaseVisionRetriever(ABC):
         **kwargs,
     ) -> Union[torch.Tensor, List[torch.Tensor]]:
         """
-        Preprocess and forward pass the passages through the model. A passage can a text chunk (e.g. BM25) or
+        Preprocess and forward pass the passages through the model. A passage can be a text chunk (e.g. BM25) or
         an image of a document page (e.g. ColPali).
 
         Args: